added some file checks

+++ install.sh
@@ -11,6 +11,9 @@ ${scrDir}/yay.sh
 # Installing packages
 ${scrDir}/install_packages.sh
 
+# Installing fonts
+fc-cache
+
 # Installing dotfiles
 mkdir -p "${HOME}/.config"
 mkdir -p "${HOME}/Pictures"
@@ -18,5 +21,11 @@ mkdir -p "${HOME}/Pictures"
 cp -r "${scrDir}/Pictures/"* "${HOME}/Pictures/"
 cp -r "${scrDir}/config/"* "${HOME}/.config/"
 
-echo "exec i3" > "${HOME}/.xinitrc"
+if [ -f "${HOME}/.xinitrc" ]; then
+    echo -e "${YELLOW}Warning: .xinitrc already exists, appending...${NC}"
+    echo "exec i3" >> "${HOME}/.xinitrc"
+else
+    echo -e "${YELLOW}Creating .xinitrc...${NC}"
+    echo "exec i3" > "${HOME}/.xinitrc"
+fi
 startx
